+++ components/Cell/Cell.scss
@@ -38,7 +38,6 @@
   &_ft {
     display: flex;
     align-items: center;
-    padding-right: pxToRem(30);
   }
   &:after {
     content: '';
