+++ components/Form/CheckBox.js
@@ -4,54 +4,86 @@ import Icon from '../Icon'
 
 const PropTypes = React.PropTypes
 // TODO 待完成,解决嵌套在 Form 组件后 checkbox 的选择问题
-export default class Checkbox extends React.Component {
+export default class _FieldCheckbox extends React.Component {
+
+  constructor(props) {
+    super(props)
+    this.state = {
+      checked: this.props.checked || false
+    }
+  }
 
   static propTypes = {
     name: PropTypes.string,
-    shouldRsa: PropTypes.bool,
-    isError: PropTypes.bool,
-    disabled: PropTypes.bool,
     required: PropTypes.bool,
+    disabled: PropTypes.bool,
     onChange: PropTypes.func,
-    formCellChange: PropTypes.func
+    handleFieldChange: PropTypes.func
   }
 
   static defaultProps = {
-    shouldRsa: false,
-    isError: false,
-    disabled: false,
     required: true,
-    onChange: () => {},
-    formCellChange: () => {}
+    disabled: false,
+    onChange: () => {
+    },
+    handleFieldChange: () => {
+    },
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.checked && nextProps.checked !== this.state.checked) {
+      this.setState({
+        checked: nextProps.checked
+      })
+    }
+  }
+
+  componentDidUpdate(preProps, preState) {
+    const {handleFieldChange} = preProps
+    if (preState.checked !== this.state.checked) {
+      handleFieldChange(this.data)
+      console.log(this.data)
+    }
+  }
+
+  get data() {
+    const {checked} = this.state
+    const {name} = this.props
+    return {
+      name,
+      value: checked
+    }
+  }
+
+  handleChange = (e) => {
+    this.props.onChange(e)
+    this.setState({
+      checked: e.target.checked
+    })
   }
 
   render() {
     const {
-      className, checked, name,
-      formCellChange, onChange,
-      value,
-      shouldRsa, validate, errorMsg, required, isError,
-      ...others
+      className, name, value, children, onChange, handleFieldChange, ...others
     } = this.props
+    const {checked} = this.state
     const cls = classNames({
       NEUI_checkbox: true,
       NEUI_checkbox_checked: checked,
       [className]: className
     })
     return (
-      <label className={cls} {...others}>
+      <div className={cls}>
         <input type="checkbox"
                name={name}
-               onChange={(e) => {
-                 console.log(e.target.checked, e.target.value)
-                 formCellChange(e.target, 'checkbox')
-                 onChange(e)
-               }}
+               onChange={this.handleChange}
                checked={checked}
                value={value}
+               {...others}
         />
         {checked ? <Icon type="tick" className="NEUI_checkbox_icon"/> : null}
-      </label>
+        {children}
+      </div>
     )
   }
 }
