+++ components/Form/Form.scss
@@ -4,7 +4,7 @@
   flex: 1;
   display: flex;
   align-items: center;
-  justify-content: flex-start;
+  justify-content: space-between;
 }
 
 .NEUI_cells_form {
@@ -19,10 +19,15 @@
 
 .NEUI_input {
   @include subCell();
-  //margin-left: pxToRem(22);
   input {
     width: 100%;
     color: $contentColor;
+    border: 0;
+    outline: 0;
+    -webkit-appearance: none;
+    &:focus {
+      outline: #000;
+    }
   }
 }
 
@@ -64,6 +69,10 @@
   }
   input {
     position: absolute;
+    /**
+    设置 z-index, 为了解决 Icon 覆盖在 checkbox 上导致无法点击到 checkbox 的问题。
+     */
+    z-index: 1;
     width: 100%;
     height: 100%;
     top: 0;
